Updated animations and styling in App.js and App.css
Modified the App.js and App.css to enhance the user experience. Removed the ScrollTrigger plugin and implemented custom scroll controlled animations for smoother and more precise animations. Instead of the road moving, now we have a car and stripes that animate based on the scroll.

The car animation stops at 40% of the viewport width after which the stripes begin to move, adding to the motion illusion. Imported new images for car and background, and updated the CSS to reflect these changes.

In App.css, introduced custom font, modified styling of elements and added new #grass and #car styles.