Adjust CSS for the responsive layout
In this commit, the CSS transformations and the width of specific CSS classes are modified in interface details to improve responsiveness and provide a much better user interface. This change was necessary to fix an alignment issue due to a design flaw in the layout. Additionally, in the App.js, a new method animateMileSignText has been added to animate miles sign based on scroll activity for better user interaction.