Add custom text and date to scroll animation
This commit introduces state changes to the scrolling animation in App.js. A newState called 'textStyle' has been added to dynamically change the display size of the text. Function 'handleDisplaySign' has been updated to set the 'textStyle' according to the currentKeyPoint. Text size and line height have been added to the key points to enhance the visual experience. Similarly, styles have been updated in App.css to reflect these changes. This will allow the text size and line spacing to be custom according to the keyed point, adding richness to the visual experience.