Refine Google Maps UI and reverse movement logic
The UI controls on Google Maps are now disabled, and CSS values have been added to style the map display. This was done to minimize distractions and give more focus to the map itself. Additionally, modified the findNextPosition function to add the ability to move in the reverse direction along the path. This caters for scenarios where the path needs to be traced back. Minor modifications are also done on App.js and App.css to adjust the positioning and appearance of map and text display components.